+++ CHANGELOG.md
@@ -20,6 +20,7 @@ All notable changes to this project will be documented in this file.
 - Fix the display of ignored pings with non-zero exit status
 - Fix a race condition in the "Change Email" flow
 - Fix grouping and sorting in the text version of the report/nag emails (#679)
+- Fix the update_timeout view to record more information for downtime bookkeeping
 
 ## v2.2.1 - 2022-06-13